Fix recovery flow for accounts without filesystems
